+++ functions/.pyproject.toml
@@ -19,11 +19,31 @@ exclude = '''
 '''
 
 [tool.isort]
+multi_line_output = 3
+force_grid_wrap = 0
+profile = "black"
 line_length = 120
 skip = '.terraform'
-sections = 'FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER'
-known_third_party = 'aws_lambda_powertools,boto3,botocore,pytest,snapshottest'
-known_first_party = 'events,notify_slack'
+sections = [
+  'FUTURE',
+  'STDLIB',
+  'THIRDPARTY',
+  'FIRSTPARTY',
+  'LOCALFOLDER'
+]
+known_third_party = [
+  'aws_lambda_powertools',
+  'boto3',
+  'botocore',
+  'pytest',
+  'snapshottest'
+]
+known_first_party = [
+  'cloudwatch',
+  'guardduty',
+  'notify_slack',
+  'utilities'
+]
 indent = '    '
 
 [tool.mypy]
