+++ functions/Pipfile
@@ -30,7 +30,7 @@ halstead = "python3 -m radon hal -i 'docs,tests' ."
 typecheck = "python3 -m mypy . --ignore-missing-imports"
 lint = "python3 -m flake8 . --count --statistics --benchmark --exit-zero --config=.flake8"
 'lint:ci' = "python3 -m flake8 . --config=.flake8"
-imports = "python3 -m isort . --profile black"
+imports = "python3 -m isort ."
 format = "python3 -m black ."
 
 [pipenv]
