+++ functions/notify_slack.py
@@ -12,173 +12,18 @@ import logging
 import os
 import urllib.parse
 import urllib.request
-from enum import Enum
 from typing import Any, Dict, Optional, Union, cast
 from urllib.error import HTTPError
 
-import boto3
 from aws_lambda_powertools import Logger
 from aws_lambda_powertools.utilities.data_classes import SNSEvent, event_source
+from cloudwatch import get_slack_attachment as get_cloudwatch_slack_attachment
+from guardduty import get_slack_attachment as get_guardduty_slack_attachment
 
 logger = Logger(service="notify-slack")
 
-# Set default region if not provided
-REGION = os.environ.get("AWS_REGION", "us-east-1")
-
-# Create client so its cached/frozen between invocations
-KMS_CLIENT = boto3.client("kms", region_name=REGION)
-
 # https://awslabs.github.io/aws-lambda-powertools-python/latest/core/logger/#logging-incoming-event
 LOG_EVENTS = os.environ.get("LOG_EVENTS", "False") == "True"
-
-
-class AwsService(Enum):
-    """AWS service supported by function"""
-
-    cloudwatch = "cloudwatch"
-    guardduty = "guardduty"
-
-
-def get_service_url(region: str, service: str) -> str:
-    """Get the appropriate service URL for the region
-
-    :param region: name of the AWS region
-    :param service: name of the AWS service
-    :returns: AWS console url formatted for the region and service provided
-    """
-    try:
-        service_name = AwsService[service].value
-        if region.startswith("us-gov-"):
-            return f"https://console.amazonaws-us-gov.com/{service_name}/home?region={region}"
-        else:
-            return f"https://console.aws.amazon.com/{service_name}/home?region={region}"
-
-    except KeyError:
-        print(f"Service {service} is currently not supported")
-        raise
-
-
-class CloudWatchAlarmState(Enum):
-    """Maps CloudWatch notification state to Slack message format color"""
-
-    OK = "good"
-    INSUFFICIENT_DATA = "warning"
-    ALARM = "danger"
-
-
-def format_cloudwatch_alarm(message: Dict[str, Any], region: str) -> Dict[str, Any]:
-    """Format CloudWatch alarm event into Slack message format
-
-    :params message: SNS message body containing CloudWatch alarm event
-    :region: AWS region where the event originated from
-    :returns: formatted Slack message payload
-    """
-
-    cloudwatch_url = get_service_url(region=region, service="cloudwatch")
-    alarm_name = message["AlarmName"]
-
-    return {
-        "color": CloudWatchAlarmState[message["NewStateValue"]].value,
-        "fallback": f"Alarm {alarm_name} triggered",
-        "fields": [
-            {"title": "Alarm Name", "value": f"`{alarm_name}`", "short": True},
-            {
-                "title": "Alarm Description",
-                "value": f"`{message['AlarmDescription']}`",
-                "short": False,
-            },
-            {
-                "title": "Alarm reason",
-                "value": f"`{message['NewStateReason']}`",
-                "short": False,
-            },
-            {
-                "title": "Old State",
-                "value": f"`{message['OldStateValue']}`",
-                "short": True,
-            },
-            {
-                "title": "Current State",
-                "value": f"`{message['NewStateValue']}`",
-                "short": True,
-            },
-            {
-                "title": "Link to Alarm",
-                "value": f"{cloudwatch_url}#alarm:alarmFilter=ANY;name={urllib.parse.quote(alarm_name)}",
-                "short": False,
-            },
-        ],
-        "text": f"AWS CloudWatch notification - {message['AlarmName']}",
-    }
-
-
-class GuardDutyFindingSeverity(Enum):
-    """Maps GuardDuty finding severity to Slack message format color"""
-
-    Low = "#777777"
-    Medium = "warning"
-    High = "danger"
-
-
-def format_guardduty_finding(message: Dict[str, Any], region: str) -> Dict[str, Any]:
-    """
-    Format GuardDuty finding event into Slack message format
-
-    :params message: SNS message body containing GuardDuty finding event
-    :params region: AWS region where the event originated from
-    :returns: formatted Slack message payload
-    """
-
-    guardduty_url = get_service_url(region=region, service="guardduty")
-    detail = message["detail"]
-    service = detail.get("service", {})
-    severity_score = detail.get("severity")
-
-    if severity_score < 4.0:
-        severity = "Low"
-    elif severity_score < 7.0:
-        severity = "Medium"
-    else:
-        severity = "High"
-
-    return {
-        "color": GuardDutyFindingSeverity[severity].value,
-        "fallback": f"GuardDuty Finding: {detail.get('title')}",
-        "fields": [
-            {
-                "title": "Description",
-                "value": f"`{detail['description']}`",
-                "short": False,
-            },
-            {
-                "title": "Finding Type",
-                "value": f"`{detail['type']}`",
-                "short": False,
-            },
-            {
-                "title": "First Seen",
-                "value": f"`{service['eventFirstSeen']}`",
-                "short": True,
-            },
-            {
-                "title": "Last Seen",
-                "value": f"`{service['eventLastSeen']}`",
-                "short": True,
-            },
-            {"title": "Severity", "value": f"`{severity}`", "short": True},
-            {
-                "title": "Count",
-                "value": f"`{service['count']}`",
-                "short": True,
-            },
-            {
-                "title": "Link to Finding",
-                "value": f"{guardduty_url}#/findings?search=id%3D{detail['id']}",
-                "short": False,
-            },
-        ],
-        "text": f"AWS GuardDuty Finding - {detail.get('title')}",
-    }
 
 
 def format_default(
@@ -244,16 +89,14 @@ def get_slack_message_payload(
     message = cast(Dict[str, Any], message)
 
     if "AlarmName" in message:
-        notification = format_cloudwatch_alarm(message=message, region=region)
-        attachment = notification
+        attachment = get_cloudwatch_slack_attachment(message=message, region=region)
 
     elif (
         isinstance(message, Dict) and message.get("detail-type") == "GuardDuty Finding"
     ):
-        notification = format_guardduty_finding(
+        attachment = get_guardduty_slack_attachment(
             message=message, region=message["region"]
         )
-        attachment = notification
 
     elif "attachments" in message or "text" in message:
         payload = {**payload, **message}
